+++ web/src/components/TopicSearchModal.jsx
@@ -11,12 +11,10 @@ import {
   TextField,
   Typography,
   Grid,
-  Switch,
   FormControl,
   Select,
   MenuItem,
 } from "@mui/material";
-import { styled } from "@mui/material/styles";
 import { DateTimePicker, LocalizationProvider } from "@mui/x-date-pickers";
 import { AdapterDateFns } from "@mui/x-date-pickers/AdapterDateFns";
 import { addDays } from "date-fns";
@@ -24,6 +22,8 @@ import PropTypes from "prop-types";
 import React, { useState } from "react";
 
 import dialogStyle from "../cssModule/dialog.module.css";
+
+import Android12Switch from "./Android12Switch";
 
 export function TopicSearchModal(props) {
   const { show, onSearch, onCancel } = props;
@@ -96,39 +96,6 @@ export function TopicSearchModal(props) {
     }
     setDateFormList(event.target.value);
   };
-
-  const Android12Switch = styled(Switch)(({ theme }) => ({
-    padding: 8,
-    "& .MuiSwitch-track": {
-      borderRadius: 22 / 2,
-      "&:before, &:after": {
-        content: "''",
-        position: "absolute",
-        top: "50%",
-        transform: "translateY(-50%)",
-        width: 16,
-        height: 16,
-      },
-      "&:before": {
-        backgroundImage: `url("data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" height="16" width="16" viewBox="0 0 24 24"><path fill="${encodeURIComponent(
-          theme.palette.getContrastText(theme.palette.primary.main),
-        )}" d="M21,7L9,19L3.5,13.5L4.91,12.09L9,16.17L19.59,5.59L21,7Z"/></svg>")`,
-        left: 12,
-      },
-      "&:after": {
-        backgroundImage: `url("data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" height="16" width="16" viewBox="0 0 24 24"><path fill="${encodeURIComponent(
-          theme.palette.getContrastText(theme.palette.primary.main),
-        )}" d="M19,13H5V11H19V13Z" /></svg>")`,
-        right: 12,
-      },
-    },
-    "& .MuiSwitch-thumb": {
-      boxShadow: "none",
-      width: 16,
-      height: 16,
-      margin: 2,
-    },
-  }));
 
   const titleForm = (
     <Grid container sx={{ margin: 1.5, width: "100%" }}>
